+++ CHANGELOG.md
@@ -32,6 +32,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Task `Build_ModuleOutput_ModuleBuilder`
   - Proper support for DSC composite resources (*.schema.psm1).
+- Added task `Set_PSModulePath`.
+  - Added function `Set-SamplerPSModulePath`.
+  - Added tests for the task and function.
+  - Added task `Set_PSModulePath` to `build.yml` Plaster template for project
+    type `dsccommunity`.
 
 ### Changed